#32: Check assets dir from every fs::load call
